+++ test/src/test_utils.js
@@ -3,8 +3,10 @@ import TransportNodeHid from "@ledgerhq/hw-transport-node-hid";
 
 import Ada, { utils, cardano } from "../../lib/Ada";
 
+export const Assert = utils.Assert;
+
 export const serializeAddressInfo = cardano.serializeAddressInfo;
-export const str_to_path = cardano.str_to_path;
+export const str_to_path = utils.str_to_path;
 export const hex_to_buf = utils.hex_to_buf;
 export const pathToBuffer = str => utils.path_to_buf(str_to_path(str));
 export const uint32_to_buf = utils.uint32_to_buf
